+++ src/mqtt.h
@@ -87,12 +87,12 @@ class Mqtt {
                                          const char * const ** options,
                                          uint8_t               options_size,
                                          const int16_t         dv_set_min,
-                                         const int16_t         dv_set_max,
+                                         const uint32_t        dv_set_max,
                                          const int8_t          num_op,
                                          const JsonObject &    dev_json);
 
     static bool publish_system_ha_sensor_config(uint8_t type, const char * name, const char * entity, const uint8_t uom);
-    static bool publish_ha_climate_config(const uint8_t tag, const bool has_roomtemp, const bool remove = false, const int16_t min = 5, const uint16_t max = 30);
+    static bool publish_ha_climate_config(const uint8_t tag, const bool has_roomtemp, const bool remove = false, const int16_t min = 5, const uint32_t max = 30);
 
     static void show_topic_handlers(uuid::console::Shell & shell, const uint8_t device_type);
     static void show_mqtt(uuid::console::Shell & shell);
